去除循环依赖  所有项目可运行

+++ part1/payment-web/handler/wrapper.go
@@ -31,7 +31,7 @@ func AuthWrapper(h http.Handler) http.Handler {
 				userId := sess.Values["userId"].(int64)
 				if userId != 0 {
 					rsp, err := authClient.GetCachedAccessToken(context.TODO(), &auth.Request{
-						UserId: uint64(userId),
+						UserId: userId,
 					})
 					if err != nil {
 						log.Logf("[AuthWrapper]，err：%s", err)
